fix: restore canonical validation tests within backend size limits
Keep pkg/config at 15 files by moving rollback coverage into
config_validator_test.go while canonical structural tests stay in a
dedicated test file under the 1000-line backend-size cap.
